+++ src/main/java/todo.txt
@@ -1,2 +1,3 @@
 
-T/0/ 2
+D/0/ sleep /2020-01-12
+E/0/ drink /2019-12-01
